Update the devices
