+++ src/app/devices/devices.component.html
@@ -13,14 +13,13 @@
         <span class="text-info outputs">{{device.Outputs.length}}<sup>Outputs</sup></span>
       </div>
     </div>
-    <hr>
     
     <!-- Input managing -->
-    <div class='flex-container vertical-container content-padding' *ngIf='device.Inputs.length'>
+    <div class='flex-container vertical-container content-padding border-top' *ngIf='device.Inputs.length'>
         <div class="vertical-row">
           <h3>Inputs</h3>
         </div>
-        <div class="vertical-row">
+        <div class="vertical-row controlls">
           <span class="button-container" *ngFor="let pin of device.Inputs">
             <label><input type="checkbox" class="ios-switch smallswitch" [attr.checked]="pin.value == 1 ? '' : null" /><div><div></div></div></label>
           </span>
@@ -28,13 +27,13 @@
         {{device.DescribeInputs()}}
     </div>
 
-    <hr>
+    
     <!-- Outputs managing -->
-    <div class='flex-container vertical-container content-padding' *ngIf='device.Outputs.length'>
+    <div class='flex-container vertical-container content-padding border-top' *ngIf='device.Outputs.length'>
         <div class="vertical-row">
           <h3>Outputs</h3>
         </div>
-        <div class="vertical-row">
+        <div class="vertical-row controlls">
           <span class="button-container" *ngFor="let pin of device.Outputs">
             <label><input type="checkbox" class="ios-switch smallswitch" [attr.checked]="pin.value == 1 ? '' : null" /><div><div></div></div></label>
           </span>
